Refactor: move occurrence API classes into occurrence from wastd/api.py.
